+++ libvast/vast/defaults.hpp
@@ -35,15 +35,6 @@ constexpr size_t max_recursion = 100;
 
 // -- constants for the import command and its subcommands ---------------------
 
-/// Contains constants that are shared by two or more import subcommands.
-namespace import::shared {
-
-/// Path for reading input events or `-` for reading from STDIN.
-constexpr std::string_view read = "-";
-
-} // namespace import::shared
-
-/// Contains constants for the import command.
 namespace import {
 
 /// Maximum size for sources that generate table slices.
@@ -72,32 +63,11 @@ constexpr std::chrono::milliseconds batch_timeout = std::chrono::seconds{10};
 constexpr std::chrono::milliseconds read_timeout
   = std::chrono::milliseconds{20};
 
-/// Contains settings for the zeek subcommand.
-struct zeek {
-  /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "vast.import.zeek";
-
-  /// Path for reading input events.
-  static constexpr auto read = shared::read;
-};
-
-/// Contains settings for the zeek-json subcommand.
-struct zeek_json {
-  /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "vast.import.zeek-json";
-
-  /// Path for reading input events.
-  static constexpr auto read = shared::read;
-};
+/// Path for reading input events or `-` for reading from STDIN.
+constexpr std::string_view read = "-";
 
 /// Contains settings for the csv subcommand.
 struct csv {
-  /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "vast.import.csv";
-
-  /// Path for reading input events.
-  static constexpr auto read = shared::read;
-
   static constexpr char separator = ',';
 
   // TODO: agree on reasonable values
@@ -106,41 +76,8 @@ struct csv {
   static constexpr std::string_view kvp_separator = "=";
 };
 
-/// Contains settings for the json subcommand.
-struct json {
-  /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "vast.import.json";
-
-  /// Path for reading input events.
-  static constexpr auto read = shared::read;
-};
-
-/// Contains settings for the suricata subcommand.
-struct suricata {
-  /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "vast.import.suricata";
-
-  /// Path for reading input events.
-  static constexpr auto read = shared::read;
-};
-
-/// Contains settings for the syslog subcommand.
-struct syslog {
-  /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "vast.import.syslog";
-
-  /// Path for reading input events.
-  static constexpr auto read = shared::read;
-};
-
 /// Contains settings for the test subcommand.
 struct test {
-  /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "vast.import.test";
-
-  /// Path for reading input events.
-  static constexpr auto read = shared::read;
-
   /// @returns a user-defined seed if available, a randomly generated seed
   /// otherwise.
   static size_t seed(const caf::settings& options);
@@ -148,12 +85,6 @@ struct test {
 
 /// Contains settings for the pcap subcommand.
 struct pcap {
-  /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "vast.import.pcap";
-
-  /// Path for reading input events.
-  static constexpr auto read = shared::read;
-
   /// Number of bytes to keep per event.
   static constexpr size_t cutoff = std::numeric_limits<size_t>::max();
 
@@ -204,15 +135,6 @@ constexpr size_t max_events_context = 100;
 // Unfortunately, `export` is a reserved keyword. The trailing `_` exists only
 // for disambiguation.
 
-/// Contains constants that are shared by two or more export subcommands.
-namespace export_::shared {
-
-/// Path for writing query results or `-` for writing to STDOUT.
-constexpr std::string_view write = "-";
-
-} // namespace export_::shared
-
-/// Contains constants for the export command.
 namespace export_ {
 
 /// Path for reading reading the query or `-` for reading from STDIN.
@@ -221,71 +143,19 @@ constexpr std::string_view read = "-";
 /// Maximum number of results.
 constexpr size_t max_events = 0;
 
-/// Contains settings for the zeek subcommand.
-struct zeek {
-  /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "vast.export.zeek";
-
-  /// Path for writing query results.
-  static constexpr auto write = shared::write;
-};
+/// Path for writing query results or `-` for writing to STDOUT.
+constexpr std::string_view write = "-";
 
 /// Contains settings for the csv subcommand.
 struct csv {
-  /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "vast.export.csv";
-
-  /// Path for writing query results.
-  static constexpr auto write = shared::write;
-
   static constexpr char separator = ',';
 
   // TODO: agree on reasonable values
   static constexpr std::string_view set_separator = " | ";
 };
 
-/// Contains settings for the ascii subcommand.
-struct ascii {
-  /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "vast.export.ascii";
-
-  /// Path for writing query results.
-  static constexpr auto write = shared::write;
-};
-
-/// Contains settings for the json subcommand.
-struct json {
-  /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "vast.export.json";
-
-  /// Path for writing query results.
-  static constexpr auto write = shared::write;
-};
-
-/// Contains settings for the null subcommand.
-struct null {
-  /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "vast.export.null";
-
-  /// Path for writing query results.
-  static constexpr auto write = shared::write;
-};
-
-struct arrow {
-  /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "vast.export.arrow";
-  /// Path for writing query results.
-  static constexpr auto write = vast::defaults::export_::shared::write;
-};
-
 /// Contains settings for the pcap subcommand.
 struct pcap {
-  /// Nested category in config files for this subcommand.
-  static constexpr const char* category = "vast.export.pcap";
-
-  /// Path for writing query results.
-  static constexpr auto write = shared::write;
-
   /// Flush to disk after that many packets.
   static constexpr size_t flush_interval = 10'000;
 };
@@ -300,7 +170,7 @@ struct infer {
   static constexpr const char* category = "vast.infer";
 
   /// Path for reading input events.
-  static constexpr auto read = defaults::import::shared::read;
+  static constexpr auto read = defaults::import::read;
 
   /// Number of bytes to buffer from input.
   static constexpr size_t buffer_size = 8'192;