+++ libvast/vast/format/arrow.hpp
@@ -32,8 +32,6 @@ namespace vast::format::arrow {
 /// An Arrow writer.
 class writer : public format::writer {
 public:
-  using defaults = vast::defaults::export_::arrow;
-
   using output_stream_ptr = std::shared_ptr<::arrow::io::OutputStream>;
 
   using batch_writer_ptr = std::shared_ptr<::arrow::ipc::RecordBatchWriter>;